+++ examples/nodejs/proofs/verifiable-presentation-id-cred.ts
@@ -120,21 +120,27 @@ const [sender, signer] = parseKeysFile(walletFile);
 // First we generate the verification request.
 //
 // This will normally happen server-side.
-const context = VerificationRequestV1.createSimpleContext(
+const requestContext = VerificationRequestV1.createSimpleContext(
     sha256([Buffer.from(Date.now().toString())]),
     randomUUID(),
     'Example VP'
 );
-const statements = VerificationRequestV1.statementBuilder()
-    .addAccountOrIdentityStatement([new IdentityProviderDID(network, identityProviderIndex)], (b) => {
+const requestClaims = VerificationRequestV1.claimsBuilder()
+    .addAccountOrIdentityClaims([new IdentityProviderDID(network, identityProviderIndex)], (b) => {
         b.addEUResidency();
         b.addMinimumAge(18);
         b.revealAttribute('firstName');
     })
-    .getStatements();
-const verificationRequest = await VerificationRequestV1.createAndAnchor(grpc, { sender, signer }, context, statements, {
-    info: 'Example VP anchor',
-});
+    .getClaims();
+const verificationRequest = await VerificationRequestV1.createAndAnchor(
+    grpc,
+    { sender, signer },
+    requestContext,
+    requestClaims,
+    {
+        info: 'Example VP anchor',
+    }
+);
 
 // simulate sending a response to the client requesting the verification request
 const requestJson = JSONBig.stringify(verificationRequest);
@@ -166,11 +172,11 @@ const requestParsed = VerificationRequestV1.fromJSON(JSONBig.parse(requestJson))
 // From the above, we retreive the secret input which is at the core of creating the verifiable presentation (proof)
 const proofInput = createIdentityCommitmentInputWithHdWallet(idObject, idp, identityIndex, wallet);
 
-// we select the identity to prove the statement for
-const idStatement = requestParsed.credentialStatements.find(
+// we select the identity to prove the claims for
+const idClaims = requestParsed.subjectClaims.find(
     (s) => s.type === 'identity'
-)! as VerificationRequestV1.IdentityStatement; // we unwrap here, as we know the statement exists (we created it just above)
-const claims = VerifiablePresentationV1.createIdentityClaims(network, idp.ipInfo.ipIdentity, idStatement.statement);
+)! as VerificationRequestV1.IdentityClaims; // we unwrap here, as we know the claims exists (we created it just above)
+const proofClaims = VerifiablePresentationV1.createIdentityClaims(network, idp.ipInfo.ipIdentity, idClaims.statements);
 
 console.log(
     'Waiting for verification request anchor transaction to finalize:',
@@ -184,7 +190,7 @@ console.log('Verification request anchor successfully registered.');
 const presentation = await VerifiablePresentationV1.createFromAnchor(
     grpc,
     requestParsed,
-    [claims],
+    [proofClaims],
     [proofInput],
     [{ label: 'ResourceID', context: 'Example VP use-case' }]
 );
@@ -196,24 +202,19 @@ const presentationJson = JSONBig.stringify(presentation);
 const presentationParsed = VerifiablePresentationV1.fromJSON(JSONBig.parse(presentationJson));
 
 // Finally, the entity requesting the proof stores the audit report and registers a pulic version on chain
-const result = await VerificationAuditRecordV1.createChecked(
+const registration = await VerificationAuditRecordV1.createAndAnchor(
     randomUUID(),
     verificationRequest,
     presentationParsed,
     grpc,
-    network
+    { metadata: { sender, signer } },
+    { network }
 );
 
-if (result.type === 'failed') throw new Error(`Failed to verify presentation: ${result.error}`);
-const auditTransaction = await VerificationAuditRecordV1.registerAnchor(
-    result.result,
-    grpc,
-    { sender, signer },
-    {
-        info: 'Some public info',
-    }
-);
+if (registration.type === 'failed') throw new Error(`Failed to verify presentation: ${registration.error}`);
 
-console.log('Waiting for verification audit report registration to finalize:', auditTransaction.toString());
-await grpc.waitForTransactionFinalization(auditTransaction);
+const { anchorTransactionRef } = registration.result;
+
+console.log('Waiting for verification audit report registration to finalize:', anchorTransactionRef.toString());
+await grpc.waitForTransactionFinalization(anchorTransactionRef);
 console.log('Verification audit anchor successfully registered.');